Move ItemsContainer to separate folder and rename to Container
